requirements.txt
asgiref==3.11.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==6.0.4
django-allauth==65.15.0
django-cors-headers==4.9.0
django-simpleui==2026.1.13
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==20.1.0
idna==3.13
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.1
psycopg2-binary==2.9.12
pycparser==3.0
PyJWT==2.12.1
python-decouple==3.8
PyYAML==6.0.3
requests==2.33.1
requests_pkcs12==1.27
setuptools==82.0.1
sqlparse==0.5.5
swagger-ui-py==25.7.1
urllib3==2.6.3
whitenoise==6.12.0
